impl: Socket-based afterfx data capture
Apparently, internally aerender.exe works by spawning the AfterFX.com process with "-s" argument that supplies a script to be ran. That script is commandLineRenderer.jsx and is located inside Ae files. Basically, we don't need to read the output of aerender.exe, but listen to Ae directly, which can potentially reduce the load on the UI thread, since we are not using dispatcher anymore :)